requirements.txt
MetaTrader5>=5.0.45
openai>=1.76.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pywin32>=306
mplfinance>=0.12.10b0
numpy>=1.24.0
pandas>=2.0.0
